#42 Improve library design
Extract block rules from bloc parser
